feat: TIO-177: rename project list file and widget
